+++ packages/create-app/template/tanstack-start/app/routes/index.tsx
@@ -13,7 +13,7 @@ function Home() {
       }}
       className="text-center py-32"
     >
-      <h1 className="font-medium text-xl mb-4">Fumadocs on Tanstack Start.</h1>
+      <h1 className="font-medium text-xl mb-4">Nopends on Tanstack Start.</h1>
       <Link
         to="/docs/$"
         className="px-3 py-2 rounded-lg bg-fd-primary text-fd-primary-foreground font-medium text-sm mx-auto"