+++ packages/typescript/src/lib/type-table.ts
@@ -63,7 +63,7 @@ export async function getTypeTableOutput(
   if (type && type.split('\n').length > 1) {
     content += `\n${type}`;
   } else if (type) {
-    typeName ??= '$Fumadocs';
+    typeName ??= '$Nopends';
     content += `\nexport type ${typeName} = ${type}`;
   }